requirements.txt
celery
selenium
pytz
django
python-dateutil
bs4
mongo
isodate
uritemplate.py==3.0.2
google-api-python-client==1.5.5
beautifulsoup4==4.5.3
requests==2.12.4
oauth2client==3.0.0
sqlalchemy
django-celery-beat
django-cache-utils
django-jsonfield>=0.6
django-nvd3>=0.5.0
django-widget-tweaks
py-moneyed 
django-money
python-memcached
